支付回退的BUG

+++ boss-core/src/main/java/com/ycsoft/business/service/impl/PayService.java
@@ -189,12 +189,20 @@ public class PayService extends BaseBusiService implements IPayService {
 					}
 					info=SystemConstants.UNPAY_CANCEL_PROMPT_FEEANDBUSI;
 				}else{
-					//产品退订，不能取消
-					throw new ServicesException(ErrorCode.UnPayOrderCancelUnsubscribe);
-					//if(!onlyShowInfo){
-					//	this.cancelUnPayProdUnsubscribePay(fee,doneCode);
-					//}
-					//info=SystemConstants.UNPAY_CANCEL_PROMPT_ONLYACCT;
+					if(StringHelper.isNotEmpty(fee.getProd_sn())){
+						//产品订单退订，不能取消
+						throw new ServicesException(ErrorCode.UnPayOrderCancelUnsubscribe);
+					}else{
+						//不对应有订单，则可以取消费用
+						if(!onlyShowInfo){
+							//作废缴费信息
+							feeComponent.saveCancelFeeUnPay(fee, doneCode);
+							//作废业务
+							doneCodeComponent.cancelDoneCode(fee.getCreate_done_code());
+						}
+						info=SystemConstants.UNPAY_CANCEL_PROMPT_ONLYFEE;
+					}
+					
 				}
 			}else{
 				if(fee.getReal_pay()>0){
@@ -325,24 +333,24 @@ public class PayService extends BaseBusiService implements IPayService {
 	private void cancelUnPayProdOrderPay(CFeeAcct fee,Integer doneCode) throws Exception{
 		
 		String order_sn=fee.getProd_sn();
-		
-		CProdOrder order=cProdOrderDao.findByKey(order_sn);
-		//检查套餐类要按订购顺序取消，同一个用户的宽带类单产品要按订购顺序取消，用户一个用户的非宽带单产品按相同产品订购顺序取消。
-		//TODO 操作过订单编辑且费用发生变化费用记录不允许取消 目的是保证c_fee_acct中pre_invalid_date和begin_date准确
-		this.checkUnPayOrderCancel(order,fee);
-		//恢复被覆盖转移的订单
-		orderComponent.recoverTransCancelOrder(order.getDone_code(),order.getCust_id(),doneCode);
-		//删除c_prod_order_fee
-		cProdOrderFeeDao.deleteOrderFeeByOrderSn(order_sn);
-		//移除订单到历史表
-		List<CProdOrder> cancelOrders=orderComponent.saveCancelProdOrder(order, doneCode);
-
+		if(StringHelper.isNotEmpty(order_sn)){
+			CProdOrder order=cProdOrderDao.findByKey(order_sn);
+			//检查套餐类要按订购顺序取消，同一个用户的宽带类单产品要按订购顺序取消，用户一个用户的非宽带单产品按相同产品订购顺序取消。
+			//TODO 操作过订单编辑且费用发生变化费用记录不允许取消 目的是保证c_fee_acct中pre_invalid_date和begin_date准确
+			this.checkUnPayOrderCancel(order,fee);
+			//恢复被覆盖转移的订单
+			orderComponent.recoverTransCancelOrder(order.getDone_code(),order.getCust_id(),doneCode);
+			//删除c_prod_order_fee
+			cProdOrderFeeDao.deleteOrderFeeByOrderSn(order_sn);
+			//移除订单到历史表
+			List<CProdOrder> cancelOrders=orderComponent.saveCancelProdOrder(order, doneCode);
+			//取消授权
+			sendProdAtuh(cancelOrders,fee.getCust_id(),doneCode);
+		}
 		//作废缴费信息
 		feeComponent.saveCancelFeeUnPay(fee, doneCode);
 		//作废业务
 		doneCodeComponent.cancelDoneCode(fee.getCreate_done_code());
-		//取消授权
-		sendProdAtuh(cancelOrders,fee.getCust_id(),doneCode);
 	}
 	/**
 	 * 加减授权都能正确处理
@@ -611,7 +619,8 @@ public class PayService extends BaseBusiService implements IPayService {
 			if(invoice==null){//发票不存在
 				 throw new ServicesException(ErrorCode.InvoiceNotExists);
 			}
-			if(!optr_id.equals(invoice.getOptr_id())){//发票不是支付人的
+			if(!BusiCodeConstants.SUPER_CANCEL_PAY.equals(this.getBusiParam().getBusiCode())
+					&&!optr_id.equals(invoice.getOptr_id())){//发票不是支付人的
 				 throw new ServicesException(ErrorCode.InvoiceIsNotYou);
 			}
 			if(!StatusConstants.USE.equals(invoice.getStatus())){//发票不是使用状态
@@ -651,15 +660,6 @@ public class PayService extends BaseBusiService implements IPayService {
 		if(invoiceIds!=null&&invoiceIds.length>0){
 			for(String invoice_id:invoiceIds){
 				if(StringHelper.isNotEmpty(invoice_id)){
-					RInvoice invoice= invoiceComponent.queryById(invoice_id, SystemConstants.BASE_INVOICE_CODE);
-					if(!BusiCodeConstants.SUPER_CANCEL_PAY.equals(busiCode)
-							&&!optrId.equals(invoice.getOptr_id())){
-						throw new ServicesException(ErrorCode.InvoiceIsNotYou);
-					}
-					if(!StatusConstants.IDLE.equals(invoice.getFinance_status())){
-						throw new ServicesException(ErrorCode.InvoiceCheckStatusIsNotIdle);
-					}
-					invoiceComponent.invalidInvoiceAndClearFeeInfo(doneCode, invoice_id, SystemConstants.BASE_INVOICE_CODE);
 					invalidInvoiceByinvoiceId(invoice_id,SystemConstants.BASE_INVOICE_CODE,doneCode);
 				}
 			}				
@@ -1065,14 +1065,16 @@ public class PayService extends BaseBusiService implements IPayService {
 				continue;
 			}
 			CFeePay pay=feeComponent.queryCFeePayByPaySn(cfee.getPay_sn());
-			if(!optrId.equals(pay.getOptr_id())){
+			if(!BusiCodeConstants.SUPER_CANCEL_PAY.equals(this.getBusiParam().getBusiCode())
+					&&!optrId.equals(pay.getOptr_id())){
 				throw new ServicesException(ErrorCode.InvoiceIsNotYou);
 			}
 			payMap.put(cfee.getPay_sn(), pay);
 		}
 		//检查发票状态是否缴销
 		RInvoice invoice= invoiceComponent.queryById(invoice_id, invoice_code);
-		if(!optrId.equals(invoice.getOptr_id())){
+		if(!BusiCodeConstants.SUPER_CANCEL_PAY.equals(this.getBusiParam().getBusiCode())
+				&&!optrId.equals(invoice.getOptr_id())){
 			throw new ServicesException(ErrorCode.InvoiceIsNotYou);
 		}
 		if(!StatusConstants.IDLE.equals(invoice.getFinance_status())){
